maint: refactor retriever + ranker
