+++ pyproject.toml
@@ -38,7 +38,7 @@ all = [
   "numpy",
   "pillow",
   "rich",
-  "scikit-learn",
+  "scikit-learn==1.7.1",
   "typer",
 ]
 
@@ -50,5 +50,6 @@ Issues = "https://github.com/linalgo/wsd/issues"
 requires = ["setuptools >= 61.0"]
 build-backend = "setuptools.build_meta"
 
-[tool.setuptools]
-packages = ["wsd"]
+[tool.setuptools.packages.find]
+where = ["."]
+include = ["wsd*"]
