+++ wsd/models/__init__.py
@@ -1,4 +1,4 @@
 """A collection of WSD models."""
-from .baseline import JMDict, Token  # noqa: F401
-from .basic import JMDictWithPointWiseRanking  # noqa: F401
-from .gemini import JMDictGeminiRanking  # noqa: F401
+from .dictionary import *  # noqa: F401
+from .rankers import *  # noqa: F401
+from .retrievers import *  # noqa: F401
