+++ wsd/parsers/__init__.py
@@ -1,7 +1,3 @@
 """A collection of parsers for various data sources."""
-# from .jmdict import *
-# from .xlwsd import *
-from .jmdict import Entry, JMDictParser
-from .xlwsd import XLWSDParser
-
-__all__=["Entry", "JMDictParser", "XLWSDParser"]
+from .jmdict import *
+from .xlwsd import *
